# public/_doc/swagger.yaml
swagger: '2.0'
info:
  description: OpenAI API ドキュメント
  version: 0.0.1
  title: OpenAI API REST APIs
#host: urapi.herokuapp.com
basePath: /api
tags:
  - name: openai
    description: OpenAI API
schemes:
  - http
  - https
paths:
  /ping:
    get:
      tags:
        - openai
      summary: ping
      description: ping(test)
      produces:
        - application/json
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /models:
    get:
      tags:
        - openai
      summary: Models API
      description: Retrieve Models
      produces:
        - application/json
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /model/{id}:
    get:
      tags:
        - openai
      summary: Retrieve Model API
      description: Retrieve Model
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: string
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /complete:
    post:
      tags:
        - openai
      summary: Analyse with file
      description: Analyse with file
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: prompt
          in: formData
          type: string
          description: 指示プロンプト
        - name: file
          in: formData
          type: file
          description: ファイル
        - name: previous_response_id
          in: formData
          type: string
          description: 前回のレスポンスID
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
        '429':
          description: Too many requests
  /completion:
    post:
      tags:
        - openai
      summary: Analyse with file(OpenAI Compatible)
      description: Analyse with file(OpenAI Compatible)
      parameters:
        - in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/ResponseRequest"
      responses:
        '200':
          description: 成功
          schema:
            $ref: "#/definitions/OpenAIResponse"
          examples:
            application/json:
              id: "resp_xxxxxxxxx"
              object: response
              status: completed
              model: "gpt-5-mini"
              output_text: |
                This is the generated text.
        '400':
          description: エラー
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error:
                message: "Invalid request"
                type: "invalid_request_error"
        '500':
          description: OpenAI API error
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
  ResponseRequest:
    type: object
    additionalProperties: true
    required:
      - input
  OpenAIResponse:
    type: object
    description: OpenAI API Response
    properties:
      id:
        type: string
      object:
        type: string
      status:
        type: string
      model:
        type: string
      output:
        type: array
        items:
          type: object
      output_text:
        type: string
      usage:
        type: object
  ErrorResponse:
    type: object
    properties:
      error:
        type: object
        properties:
          message:
            type: string
          type:
            type: string